[RFC] Bugfix and enhancements for crate containerd-shim

+++ crates/shim/src/lib.rs
@@ -22,11 +22,13 @@ use std::error;
 use std::fs;
 use std::hash::Hasher;
 use std::io::{self, Write};
+use std::os::unix::fs::FileTypeExt;
+use std::os::unix::io::AsRawFd;
 use std::os::unix::io::RawFd;
-use std::path::PathBuf;
+use std::os::unix::net::UnixListener;
+use std::path::{Path, PathBuf};
 use std::process::{self, Command, Stdio};
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
+use std::sync::{Arc, Condvar, Mutex};
 
 pub use containerd_shim_client as protos;
 
@@ -68,6 +70,7 @@ pub struct Config {
 }
 
 /// Startup options received from containerd to start new shim instance.
+///
 /// These will be passed via [`Shim::start_shim`] to shim.
 #[derive(Debug, Default)]
 pub struct StartOpts {
@@ -84,45 +87,52 @@ pub struct StartOpts {
 }
 
 /// Helper structure that wraps atomic bool to signal shim server when to shutdown the TTRPC server.
+///
 /// Shim implementations are responsible for calling [`Self::signal`].
 #[derive(Clone)]
-pub struct ExitSignal(Arc<AtomicBool>);
+pub struct ExitSignal(Arc<(Mutex<bool>, Condvar)>);
 
 impl Default for ExitSignal {
+    #[allow(clippy::mutex_atomic)]
     fn default() -> Self {
-        ExitSignal(Arc::new(AtomicBool::new(false)))
+        ExitSignal(Arc::new((Mutex::new(false), Condvar::new())))
     }
 }
 
 impl ExitSignal {
     /// Set exit signal to shutdown shim server.
     pub fn signal(&self) {
-        self.0.store(true, Ordering::Release)
+        let (lock, cvar) = &*self.0;
+        let mut exit = lock.lock().unwrap();
+        *exit = true;
+        cvar.notify_all();
     }
 
     /// Wait for the exit signal to be set.
-    fn wait(&self) {
-        while !self.0.load(Ordering::Acquire) {
-            std::hint::spin_loop();
+    pub fn wait(&self) {
+        let (lock, cvar) = &*self.0;
+        let mut started = lock.lock().unwrap();
+        while !*started {
+            started = cvar.wait(started).unwrap();
         }
     }
 }
 
 /// Main shim interface that must be implemented by all shims.
+///
 /// Start and delete routines will be called to handle containerd's shim lifecycle requests.
-pub trait Shim: Task {
+pub trait Shim {
     /// Error type to be returned when starting/deleting shim.
     type Error: error::Error;
 
-    fn new(
-        id: &str,
-        namespace: &str,
-        publisher: RemotePublisher,
-        config: &mut Config,
-        exit: ExitSignal,
-    ) -> Self;
+    /// Type to provide task service for the shim.
+    type T: Task + Send + Sync;
+
+    /// Create a new instance of Shim.
+    fn new(id: &str, namespace: &str, publisher: RemotePublisher, config: &mut Config) -> Self;
 
     /// Start shim will be called by containerd when launching new shim instance.
+    ///
     /// It expected to return TTRPC address containerd daemon can use to communicate with
     /// the given shim instance.
     /// See https://github.com/containerd/containerd/tree/master/runtime/v2#start
@@ -132,6 +142,12 @@ pub trait Shim: Task {
     fn delete_shim(&mut self) -> Result<DeleteResponse, Self::Error> {
         Ok(DeleteResponse::default())
     }
+
+    /// Wait for the shim to exit.
+    fn wait(&mut self);
+
+    /// Get the task service object.
+    fn get_task_service(&self) -> Self::T;
 }
 
 /// Shim entry point that must be invoked from `main`.
@@ -157,15 +173,8 @@ where
     let publisher = publisher::RemotePublisher::new(&ttrpc_address)?;
 
     // Create shim instance
-    let exit_signal = ExitSignal::default();
     let mut config = Config::default();
-    let mut shim = T::new(
-        id,
-        &flags.namespace,
-        publisher,
-        &mut config,
-        exit_signal.clone(),
-    );
+    let mut shim = T::new(id, &flags.namespace, publisher, &mut config);
 
     if !config.no_sub_reaper {
         reap::set_subreaper()?;
@@ -205,8 +214,8 @@ where
                 logger::init(flags.debug)?;
             }
 
-            let task_service = create_task(Arc::new(Box::new(shim)));
-
+            let task = shim.get_task_service();
+            let task_service = create_task(Arc::new(Box::new(task)));
             let mut server = Server::new().register_service(task_service);
 
             server = if flags.socket.is_empty() {
@@ -218,7 +227,7 @@ where
             server.start()?;
 
             info!("Shim successfully started, waiting for exit signal...");
-            exit_signal.wait();
+            shim.wait();
 
             info!("Shutting down shim instance");
             server.shutdown();
@@ -275,7 +284,7 @@ pub const SOCKET_ROOT: &str = "/run/containerd";
 #[cfg(target_os = "macos")]
 pub const SOCKET_ROOT: &str = "/var/run/containerd";
 
-/// Make socket path from namespace and id.
+/// Make socket path from containerd socket path, namespace and id.
 pub fn socket_address(socket_path: &str, namespace: &str, id: &str) -> String {
     let path = PathBuf::from(socket_path)
         .join(namespace)
@@ -292,55 +301,44 @@ pub fn socket_address(socket_path: &str, namespace: &str, id: &str) -> String {
     format!("{}/{:x}.sock", SOCKET_ROOT, hash)
 }
 
-fn start_listener(address: &str) -> Result<RawFd, Error> {
-    use nix::fcntl::*;
-    use nix::sys::socket::*;
-
-    let fd = socket(
-        AddressFamily::Unix,
-        SockType::Stream,
-        SockFlag::empty(),
-        None,
-    )?;
-
-    fcntl(fd, FcntlArg::F_SETFL(OFlag::O_NONBLOCK))?;
-
-    let unix_addr = UnixAddr::new(address)?;
-
-    fs::create_dir_all(SOCKET_ROOT)?;
-
-    let socket_address = SockAddr::Unix(unix_addr);
-
-    match bind(fd, &socket_address) {
-        Ok(_) => {}
-        Err(err) if err == Errno::EADDRINUSE => {
-            fs::remove_file(address)?;
-            bind(fd, &socket_address)?;
-        }
-        Err(err) => return Err(err.into()),
+fn start_listener(address: &str) -> Result<UnixListener, Error> {
+    // Try to create the needed directory hierarchy.
+    if let Some(parent) = Path::new(address).parent() {
+        fs::create_dir_all(parent)?;
     }
 
-    listen(fd, 10)?;
-
-    Ok(fd)
+    UnixListener::bind(address).or_else(|e| {
+        if e.kind() == io::ErrorKind::AddrInUse {
+            if let Ok(md) = Path::new(address).metadata() {
+                if md.file_type().is_socket() {
+                    fs::remove_file(address)?;
+                    return UnixListener::bind(address).map_err(|e| e.into());
+                }
+            }
+        }
+        Err(e.into())
+    })
 }
 
 /// Spawn is a helper func to launch shim process.
 /// Typically this expected to be called from `StartShim`.
-pub fn spawn(opts: StartOpts) -> Result<String, Error> {
+pub fn spawn(opts: StartOpts, vars: Vec<(&str, &str)>) -> Result<String, Error> {
+    let cmd = env::current_exe()?;
+    let cwd = env::current_dir()?;
     let address = socket_address(&opts.address, &opts.namespace, &opts.id);
 
     // Create socket and prepare listener.
     // We'll use `add_listener` when creating TTRPC server.
-    let fd = start_listener(&address)?;
+    let listener = start_listener(&address)?;
+    let mut command = Command::new(cmd);
 
-    let result = Command::new(env::current_exe()?)
-        .current_dir(env::current_dir()?)
+    command
+        .current_dir(cwd)
         .stdout(Stdio::null())
         .stdin(Stdio::null())
         .stderr(Stdio::null())
         .fd_mappings(vec![FdMapping {
-            parent_fd: fd,
+            parent_fd: listener.as_raw_fd(),
             child_fd: SOCKET_FD,
         }])?
         .args(&[
@@ -351,16 +349,13 @@ pub fn spawn(opts: StartOpts) -> Result<String, Error> {
             "-address",
             &opts.address,
         ])
-        .spawn();
+        .envs(vars);
 
-    match result {
-        Ok(_) => Ok(format!("unix://{}", address)),
-        Err(err) => {
-            // Close listener if something went wrong
-            unsafe { libc::close(fd) };
-            Err(err.into())
-        }
-    }
+    command.spawn().map_err(Into::into).map(|_| {
+        // Ownership of `listener` has been passed to child.
+        std::mem::forget(listener);
+        format!("unix://{}", address)
+    })
 }
 
 #[cfg(test)]
@@ -382,5 +377,29 @@ mod tests {
         if let Err(err) = handle.join() {
             panic!("{:?}", err);
         }
+    }
+
+    #[test]
+    fn test_start_listener() {
+        let tmpdir = tempfile::tempdir().unwrap();
+        let path = tmpdir.path().to_str().unwrap().to_owned();
+
+        // A little dangerous, may be turned on under controlled environment.
+        //assert!(start_listener("/").is_err());
+        //assert!(start_listener("/tmp").is_err());
+
+        let socket = path + "/ns1/id1/socket";
+        let _listener = start_listener(&socket).unwrap();
+        let _listener2 = start_listener(&socket).unwrap();
+
+        let socket2 = socket + "/socket";
+        assert!(start_listener(&socket2).is_err());
+
+        let path = tmpdir.path().to_str().unwrap().to_owned();
+        let txt_file = path + "demo.txt";
+        fs::write(&txt_file, "test").unwrap();
+        assert!(start_listener(&txt_file).is_err());
+        let context = fs::read_to_string(&txt_file).unwrap();
+        assert_eq!(context, "test");
     }
 }
